Move route and command definitions out of the Helpers module
Alice.Router.Helpers is not the right place for these. Instead, add them
directly to the Router module. This inflates the Router module a bit but
the state handling can be moved out at a later date.
